feat: Define dynamic newState function
Create state package and relativeURL + schemeAndHost
StateFuncs. The value of SESSION_DOMAIN is used to pick
the strategy. If SESSION_DOMAIN is set then
AuthService will be able to serve requests
for all subdomains of SESSION_DOMAIN.

Also, because createNonce is not used anywhere
else it makes since to move it to this package.

This commit adds 2 new env vars:
2. SCHEME_DEFAULT - default request scheme
3. SCHEME_HEADER - header to get original request scheme from

Signed-off-by: Adam Setters <asetty@arista.com>