feat: Support host + path based routing
Create StateFunc type with relativeURL + schemeAndHost strategies for
forwarding requests. The value of SESSION_DOMAIN is used to pick the
strategy. If SESSION_DOMAIN is set then AuthService will be able to
serve requests for all subdomains of SESSION_DOMAIN and the host + path
of the request will be used when saving the oidc state for the request.

This commit adds 1 new env var:
* SCHEME_DEFAULT - default request schem (default: https)

Signed-off-by: Adam Setters <asetty@arista.com>