session: Try to load a session from header or cookie
The sessionAuthenticator can authenticate a user based on a session id.
The session id can be in two places right now:
1. A cookie
2. A header

The old logic was:
1. Check if the given header contains a value.
2. If it does, try to load a session from the value and return
   immediately.
3. Else, try to load a session from the given cookie.

The problem is in step (2). If loading the session from the header does
not work, it should also try loading the session from the cookie.

So the new logic is:
1. Check if the given header contains a value.
2. If it does, try to load a session from the header value.
3. If loading the session does not fail, use that session.
4. Else, try to load a session from the given cookie.

Github-PR: #55

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Alex Pyrgiotis <apyrgio@arrikto.com>